Converted all to be double precision.

+++ c++/layers.cpp
@@ -18,24 +18,24 @@ Dense::Dense(int n_inputs, int n_neurons):
 
 void Dense::randomize_weights() {
     std::default_random_engine generator;
-    std::uniform_real_distribution<float> distribution(0.0,1.0);
+    std::uniform_real_distribution<double> distribution(0.0,1.0);
     for (int i = 0; i < weights.rows; i++) {
 	for (int j = 0; j < weights.cols; j++) {
-	    weights[i][j] = (float) 1.0;//distribution(generator);
+	    weights[i][j] = (double) 1.0;//distribution(generator);
 	}
     }
 }
 
-matrix::Matrix<float> Dense::forward(matrix::Matrix<float> input) {
+matrix::Matrix<double> Dense::forward(matrix::Matrix<double> input) {
     // Calculate the dot product between each neuron and input data
     return matrix::add(matrix::dot(input, weights), biases);
 }
 
-matrix::Matrix<float> Dense::backward(matrix::Matrix<float> inputs, 
-	    matrix::Matrix<float> dinput) {
+matrix::Matrix<double> Dense::backward(matrix::Matrix<double> inputs, 
+	    matrix::Matrix<double> dinput) {
 
 
-    matrix::Matrix<float> temp = matrix::dot(matrix::transpose(inputs), dinput);
+    matrix::Matrix<double> temp = matrix::dot(matrix::transpose(inputs), dinput);
     dweights = temp;
 
     dbiases = matrix::sum(dinput, 0, true);
@@ -43,25 +43,25 @@ matrix::Matrix<float> Dense::backward(matrix::Matrix<float> inputs,
     return matrix::dot(dinput, matrix::transpose(weights));
 }
 
-matrix::Matrix<float> Dense::get_dweights() {
+matrix::Matrix<double> Dense::get_dweights() {
     return dweights;
 }
 
-matrix::Matrix<float> Dense::get_dbiases() {
+matrix::Matrix<double> Dense::get_dbiases() {
     return dbiases;
 }
 
-matrix::Matrix<float> Dense::get_biases() {
+matrix::Matrix<double> Dense::get_biases() {
     return biases;
 }
 
-matrix::Matrix<float> Dense::get_weights() {
+matrix::Matrix<double> Dense::get_weights() {
     return weights;
 }
-void Dense::set_biases(matrix::Matrix<float> new_biases) {
+void Dense::set_biases(matrix::Matrix<double> new_biases) {
     biases = new_biases;
 }
-void Dense::set_weights(matrix::Matrix<float> new_weights) {
+void Dense::set_weights(matrix::Matrix<double> new_weights) {
     weights = new_weights;
 }
 
@@ -69,7 +69,7 @@ void Dense::set_weights(matrix::Matrix<float> new_weights) {
 
 ReLU::ReLU(): inputs(1, 1) {}
 
-matrix::Matrix<float> ReLU::forward(matrix::Matrix<float> input) {
+matrix::Matrix<double> ReLU::forward(matrix::Matrix<double> input) {
     for (int i = 0; i < input.rows; i++) {
 	for (int j = 0; j < input.cols; j++) {
 	    if (input[i][j] < 0) {
@@ -80,8 +80,8 @@ matrix::Matrix<float> ReLU::forward(matrix::Matrix<float> input) {
     return input;
 }
 
-matrix::Matrix<float> ReLU::backward(matrix::Matrix<float> inputs, 
-	    matrix::Matrix<float> dinput) {
+matrix::Matrix<double> ReLU::backward(matrix::Matrix<double> inputs, 
+	    matrix::Matrix<double> dinput) {
     for (int i = 0; i < dinput.rows; i++) {
 	for (int j = 0; j < dinput.cols; j++) {
 	    if (inputs[i][j] <= 0) {
@@ -96,13 +96,13 @@ matrix::Matrix<float> ReLU::backward(matrix::Matrix<float> inputs,
 
 Softmax::Softmax(): inputs(1, 1) {}
 
-matrix::Matrix<float> Softmax::forward(matrix::Matrix<float> input) {
-    matrix::Matrix<float> temp = matrix::exp(matrix::subtract(input, 
+matrix::Matrix<double> Softmax::forward(matrix::Matrix<double> input) {
+    matrix::Matrix<double> temp = matrix::exp(matrix::subtract(input, 
 		matrix::max(input)));
     return matrix::division(temp, matrix::sum(temp));
 }
 
-matrix::Matrix<float> Softmax::backward(matrix::Matrix<float> dinput) {
+matrix::Matrix<double> Softmax::backward(matrix::Matrix<double> dinput) {
     return dinput;
 }
 
@@ -110,24 +110,24 @@ matrix::Matrix<float> Softmax::backward(matrix::Matrix<float> dinput) {
 
 SoftmaxCrossEntropy::SoftmaxCrossEntropy(void): softmax(), crossEntropy() {}
 
-matrix::Matrix<float> SoftmaxCrossEntropy::forward(matrix::Matrix<float> input, 
-	    matrix::Matrix<float> y_true) {
-    matrix::Matrix<float> out = softmax.forward(input);
+matrix::Matrix<double> SoftmaxCrossEntropy::forward(matrix::Matrix<double> input, 
+	    matrix::Matrix<double> y_true) {
+    matrix::Matrix<double> out = softmax.forward(input);
     loss = crossEntropy.calculateLoss(out, y_true);
     return out;
 }
 
-matrix::Matrix<float> SoftmaxCrossEntropy::backward(matrix::Matrix<float> dinput, 
-	    matrix::Matrix<float> y_true) {
+matrix::Matrix<double> SoftmaxCrossEntropy::backward(matrix::Matrix<double> dinput, 
+	    matrix::Matrix<double> y_true) {
     // Expects y_true to be one hot encoded
     matrix::Matrix<int> converted = matrix::argmax(y_true);
     for (int i = 0; i < dinput.rows; i++) {
 	dinput[i][converted[i][0]] -= 1;
     }
-    matrix::Matrix<float> temp(1, 1, dinput.rows);
+    matrix::Matrix<double> temp(1, 1, dinput.rows);
     return matrix::division(dinput, temp);
 }
 
-float SoftmaxCrossEntropy::get_loss() {
+double SoftmaxCrossEntropy::get_loss() {
     return loss;
 }
